+++ core/engine.go
@@ -3,7 +3,9 @@ package core
 import (
 	"context"
 	"fmt"
+	"math"
 	"math/big"
+	"sync"
 	"time"
 
 	"github.com/sirupsen/logrus"
@@ -23,6 +25,8 @@ type Engine struct {
 	signalChan        chan *types.Signal
 	prevBlock         *types.Block
 	finalBlock        *types.Block
+	tearedDown        bool
+	teardownOnce      sync.Once
 	withSkippedBlocks bool
 	withReorgs        bool
 }
@@ -41,6 +45,8 @@ func NewEngine(genesisHash string, genesisHeight uint64, genesisTime time.Time, 
 		blockChan:         make(chan *types.Block),
 		signalChan:        make(chan *types.Signal),
 		flashBlockChan:    make(chan *types.FlashBlock),
+		tearedDown:        false,
+		teardownOnce:      sync.Once{},
 		withSkippedBlocks: withSkippedBlocks,
 		withReorgs:        withReorgs,
 	}
@@ -57,15 +63,51 @@ func (e *Engine) Initialize(prevBlock *types.Block, finalBlock *types.Block) err
 	return nil
 }
 
-func (e *Engine) StartBlockProduction(ctx context.Context, withSignal, withFlashBlocks bool) {
-	ticker := time.NewTicker(e.blockRate)
+func (e *Engine) stop(reason string, tickers ...*time.Ticker) {
+	e.teardownOnce.Do(func() {
+		logrus.Info(reason)
 
-	logrus.WithField("rate", e.blockRate).Info("starting block producer")
-	if e.stopHeight > 0 {
-		logrus.WithField("stop_height", e.stopHeight).Info("block production will stop at height")
-		if e.prevBlock != nil && e.prevBlock.Header.Height >= e.stopHeight {
+		e.tearedDown = true
+
+		close(e.blockChan)
+		close(e.signalChan)
+		close(e.flashBlockChan)
+
+		for _, ticker := range tickers {
 			ticker.Stop()
 		}
+	})
+}
+
+func (e *Engine) StartBlockProduction(ctx context.Context, withCommitmentSignal, withFlashBlocks bool) {
+	logrus.
+		WithField("genesis_burst", e.genesisBlockBurst).
+		WithField("rate", e.blockRate).
+		WithField("size", e.blockSizeInBytes).
+		WithField("stop_height", e.stopHeight).
+		Info("starting block producer")
+
+	if e.prevBlock == nil {
+		genesisBlock := types.GenesisBlock(e.genesisHash, e.genesisHeight, e.genesisTime)
+		logrus.WithField("block", blockRef{genesisBlock.Header.Hash, e.genesisHeight}).WithField("burst", e.genesisBlockBurst).Info("starting from genesis block height")
+		e.prevBlock = genesisBlock
+		e.finalBlock = genesisBlock
+
+		e.blockChan <- genesisBlock
+
+		startBurst := time.Now()
+		for i := 0; i < int(e.genesisBlockBurst); {
+			for _, block := range e.createBlocks() {
+				if e.hasReachedStopHeight(block.Header.Height) {
+					e.stop("reached stop block height during genesis burst")
+					return
+				}
+
+				e.blockChan <- block
+				i++
+			}
+		}
+		logrus.WithField("duration", time.Since(startBurst).String()).Infof("genesis block burst of %d blocks produced", int(e.genesisBlockBurst))
 	}
 
 	var lastBlock *types.Block
@@ -73,36 +115,42 @@ func (e *Engine) StartBlockProduction(ctx context.Context, withSignal, withFlash
 	var lastFlashBlockNum uint64
 	var lastFlashBlockIndex uint64
 
-	signalTicker := time.NewTicker(e.blockRate)
-	if withSignal {
+	blockTicker := time.NewTicker(e.blockRate)
+	commitmentSignalTicker := time.NewTicker(e.blockRate)
+	flashBlockTicker := time.NewTicker(e.blockRate / 5) // we use 4 slots out of 5
+
+	if withCommitmentSignal {
 		<-time.After(e.blockRate / 2) // offset by half duration
-		signalTicker.Reset(e.blockRate)
+		commitmentSignalTicker.Reset(e.blockRate)
 	} else {
-		signalTicker.Stop()
+		commitmentSignalTicker.Stop()
 	}
 
-	flashRate := e.blockRate / 5 // we use 4 slots out of 5
-	flashBlockTicker := time.NewTicker(flashRate)
 	if !withFlashBlocks {
 		flashBlockTicker.Stop()
 	}
 
 	for {
-		select {
-		case <-ticker.C:
-			for _, block := range e.createBlocks() {
-				e.blockChan <- block
-				lastBlock = block
+		if e.tearedDown {
+			logrus.Info("block producer has been stopped")
+			return
+		}
 
-				if e.stopHeight > 0 && block.Header.Height >= e.stopHeight {
-					logrus.Info("reached stop height")
-					ticker.Stop()
+		select {
+		case <-blockTicker.C:
+			for _, block := range e.createBlocks() {
+				if e.hasReachedStopHeight(block.Header.Height) {
+					e.stop("reached stop block height", blockTicker, commitmentSignalTicker, flashBlockTicker)
 					return
 				}
+
+				e.blockChan <- block
+				lastBlock = block
 			}
-		case <-signalTicker.C:
-			if !withSignal {
-				continue // just ignore if a signal ticker comes in, but it actually should not be called because of the Stop(), unless there is a crazy race condtition
+		case <-commitmentSignalTicker.C:
+			if !withCommitmentSignal {
+				// Just ignore if a signal ticker comes in, but it actually should not be called because of the Stop(), unless there is a crazy race condition
+				continue
 			}
 			if lastBlock != nil {
 				if lastSignal != nil && lastSignal.BlockID == lastBlock.Header.Hash {
@@ -119,7 +167,8 @@ func (e *Engine) StartBlockProduction(ctx context.Context, withSignal, withFlash
 
 		case <-flashBlockTicker.C:
 			if !withFlashBlocks {
-				continue // just ignore if a flashblock ticker comes in, but it actually should not be called because of the Stop(), unless there is a crazy race condtition
+				// Just ignore if a flashblock ticker comes in, but it actually should not be called because of the Stop(), unless there is a crazy race condition
+				continue
 			}
 			if lastBlock == nil {
 				continue
@@ -144,11 +193,18 @@ func (e *Engine) StartBlockProduction(ctx context.Context, withSignal, withFlash
 			lastFlashBlockNum = num
 
 		case <-ctx.Done():
-			logrus.Info("stopping block producer")
-			close(e.blockChan)
+			e.stop("context done", blockTicker, commitmentSignalTicker, flashBlockTicker)
 			return
 		}
 	}
+}
+
+func (e *Engine) hasReachedStopHeight(height uint64) bool {
+	if e.stopHeight == 0 {
+		return false
+	}
+
+	return height > e.stopHeight
 }
 
 func (e *Engine) SubscribeBlocks() <-chan *types.Block {
@@ -164,71 +220,133 @@ func (e *Engine) SubscribeFlashBlocks() <-chan *types.FlashBlock {
 }
 
 func (e *Engine) createBlocks() (out []*types.Block) {
-	if e.prevBlock == nil {
-		genesisBlock := types.GenesisBlock(e.genesisHash, e.genesisHeight, e.genesisTime)
-		logrus.WithField("block", blockRef{genesisBlock.Header.Hash, e.genesisHeight}).Info("starting from genesis block height")
-		e.prevBlock = genesisBlock
-		e.finalBlock = genesisBlock
-
-		out = append(out, genesisBlock)
-
-		for len(out)-1 < int(e.genesisBlockBurst) {
-			out = append(out, e.createBlocks()...)
-		}
-
-		return
-	}
-
 	heightToProduce := e.prevBlock.Header.Height + 1
 	if e.withSkippedBlocks && heightToProduce%13 == 0 {
 		heightToProduce += 1
-		logrus.Info(fmt.Sprintf("skipping block #%d that is a multiple of 13, producing %d instead", heightToProduce-1, heightToProduce))
+		logrus.Info(fmt.Sprintf("skipping block #%d that is a multiple of 13, created %d instead", heightToProduce-1, heightToProduce))
 	}
 
 	if e.withReorgs && heightToProduce%17 == 0 {
 		if heightToProduce%2 == 0 {
-			logrus.Info("producing 2 block fork sequence")
+			logrus.Info("created 2 block fork sequence")
 			firstFork := e.newBlock(heightToProduce, ptr(uint64(1)), e.prevBlock)
 			secondFork := e.newBlock(heightToProduce+1, ptr(uint64(2)), firstFork)
 
 			out = append(out, firstFork, secondFork)
 		} else {
-			logrus.Info("producing 1 block fork sequence")
+			logrus.Info("created 1 block fork sequence")
 			out = append(out, e.newBlock(heightToProduce, ptr(uint64(1)), e.prevBlock))
 		}
 	}
 
 	block := e.newBlock(heightToProduce, nil, e.prevBlock)
-
 	e.addTransactions(block, e.blockSizeInBytes)
 
 	out = append(out, block)
 
 	e.prevBlock = block
 	if block.Header.Height%10 == 0 {
-		logrus.WithField("block", blockRef{block.Header.Hash, block.Header.Height}).Info("produced block is now the final block")
+		logrus.WithField("block", blockRef{block.Header.Hash, block.Header.Height}).Info("created block is now the final block")
 		e.finalBlock = block
 	}
 
 	return
 }
 
-func (e *Engine) addTransactions(block *types.Block, sizeInBytes int) {
+var simulateTypes = []string{"transfer", "delegate", "undelegate", "reward", "slash"}
+var bigZero = big.NewInt(0)
 
-	for size := block.ApproximatedSize(); size < sizeInBytes; size = block.ApproximatedSize() {
+const (
+	KiB = 1024
+)
+
+func (e *Engine) addTransactions(block *types.Block, sizeInBytes int) {
+	addTx := func(data []byte) {
 		i := len(block.Transactions)
 
+		txHash := types.MakeFakeHash(block.Header.Height, i)
+		sender := "0x" + txHash[:40]
+		receiver := "0x" + txHash[24:64]
+		amount := new(big.Int).SetUint64((block.Header.Height << 32) | uint64(i))
+		success := true
+
+		// Each five transactions, make a fixed sender
+		if i%7 == 0 {
+			sender = "0xDEADBEEF"
+		}
+
+		// Each eleven transactions, make a fixed receiver
+		if i%11 == 0 {
+			receiver = "0xBAAAAAAD"
+		}
+
+		// Each 3 transactions, make amount zero
+		if i%3 == 0 {
+			amount = bigZero
+		}
+
+		// Each 13 transactions, make it fail
+		if i%13 == 0 {
+			success = false
+		}
+
 		block.Transactions = append(block.Transactions, types.Transaction{
-			Type:     "transfer",
-			Hash:     types.MakeHash(fmt.Sprintf("%v-%v", block.Header.Height, i)),
-			Sender:   "0xDEADBEEF",
-			Receiver: "0xBAAAAAAD",
-			Amount:   big.NewInt(int64(i * 1000000000)),
-			Fee:      big.NewInt(10000),
-			Success:  true,
+			Type:     simulateTypes[i%len(simulateTypes)],
+			Hash:     txHash,
+			Sender:   sender,
+			Receiver: receiver,
+			Data:     fillData(data, block.Header.Height, i),
+			Amount:   amount,
+			Fee:      new(big.Int).SetUint64(block.Header.Height + uint64(i)),
+			Success:  success,
 			Events:   e.generateEvents(block.Header.Height),
 		})
 	}
+
+	// At those small size, the loop below is fast enough
+	if sizeInBytes < 10*KiB {
+		for size := block.ApproximatedSize(); size <= sizeInBytes; size = block.ApproximatedSize() {
+			addTx(make([]byte, 32))
+		}
+	}
+
+	// At larger sizes, we estimate the number of transactions to add and generate data
+	txCount := targetTxCount(sizeInBytes)
+	dataSizePerTx := (sizeInBytes / txCount) - 250 // rough estimate of tx overhead
+
+	for range txCount {
+		addTx(make([]byte, dataSizePerTx))
+	}
+}
+
+// targetTxCount provides an estimated target transaction count, and it gives roughly
+// in that range:
+//
+//   - 100 transactions at 100KiB
+//   - 1000 transactions at 1MiB
+//   - 2500 transactions at 2.5MiB
+//   - 5000 transactions at 5MiB
+//   - 10000 transactions at 10MiB
+//   - 100000 transactions at 100MiB
+//   - etc.
+//
+// With a minimum of 10 transactions.
+func targetTxCount(blockSizeInBytes int) int {
+	scale := math.Log10(float64(blockSizeInBytes))
+	exponent := scale - 3
+	if exponent < 1 {
+		exponent = 1
+	}
+
+	return int(math.Pow(10, float64(exponent)))
+}
+
+func fillData(buf []byte, blockHeight uint64, txIndex int) []byte {
+	for i := range buf {
+		buf[i] = byte((blockHeight + uint64(txIndex) + uint64(i)) % 256)
+	}
+
+	return buf
 }
 
 func (e *Engine) newBlock(height uint64, nonce *uint64, parent *types.Block) *types.Block {
@@ -249,15 +367,23 @@ func (e *Engine) newBlock(height uint64, nonce *uint64, parent *types.Block) *ty
 func (e *Engine) generateEvents(height uint64) []types.Event {
 	events := []types.Event{}
 
+	events = append(events, types.Event{
+		Type: "token_transfer",
+		Attributes: []types.Attribute{
+			{Key: "foo", Value: "bar"},
+		},
+	})
+
 	switch {
-	case height%3 == 0:
+	case height%2 == 0:
 		events = append(events, types.Event{
-			Type: "token_transfer",
+			Type: "coin_spent",
 			Attributes: []types.Attribute{
-				{Key: "foo", Value: "bar"},
+				{Key: "spender", Value: "fizz"},
+				{Key: "amount", Value: "buzz"},
 			},
 		})
-	case height%5 == 0:
+	case height%3 == 0:
 		events = append(events, types.Event{
 			Type: "delegate",
 			Attributes: []types.Attribute{
@@ -266,12 +392,12 @@ func (e *Engine) generateEvents(height uint64) []types.Event {
 				{Key: "amount", Value: "123456789"},
 			},
 		})
-	case height%3 == 0 && height%5 == 0:
+	case height%5 == 0:
 		events = append(events, types.Event{
-			Type: "coin_spent",
+			Type: "undelegate",
 			Attributes: []types.Attribute{
-				{Key: "spender", Value: "fizz"},
-				{Key: "amount", Value: "buzz"},
+				{Key: "delegator", Value: "addr1"},
+				{Key: "amount", Value: "123456789"},
 			},
 		})
 	}